pyproject.toml
[project]
name = "himitsubako"
version = "0.1.0.dev0"
description = "Multi-backend credential abstraction for solo Python developers. SOPS+age, macOS Keychain, Bitwarden CLI, direnv, env vars."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "originalrgsec" },
]
keywords = ["secrets", "credentials", "sops", "age", "direnv", "bitwarden", "keychain", "pydantic-settings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
keychain = ["keyring>=24.0"]
pydantic-settings = ["pydantic-settings>=2.0"]
dev = [
    "pytest>=8.0,!=9.0.3",
    "pytest-cov>=5.0",
    "ruff>=0.8,!=0.15.10",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[project.scripts]
hmb = "himitsubako.cli:main"

[project.urls]
Homepage = "https://github.com/originalrgsec/himitsubako"
Issues = "https://github.com/originalrgsec/himitsubako/issues"
Repository = "https://github.com/originalrgsec/himitsubako"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/himitsubako"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: fast unit tests (default)",
    "integration: tests requiring external binaries (sops, age, bw, direnv)",
    "macos: tests requiring macOS Keychain",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "SIM", "TCH", "RUF",
]

[tool.ruff.lint.isort]
known-first-party = ["himitsubako"]

[tool.coverage.run]
source = ["src/himitsubako"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
